requirements.txt
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
certifi==2024.12.14
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements.in
    #   requests
dj-database-url==2.3.0
    # via
    #   -r requirements.in
    #   task-manager
django==5.1.5
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-bootstrap5
    #   django-filter
    #   djangorestframework
    #   task-manager
django-bootstrap5==24.3
    # via
    #   -r requirements.in
    #   task-manager
django-filter==24.3
    # via
    #   -r requirements.in
    #   task-manager
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   task-manager
gunicorn==23.0.0
    # via
    #   -r requirements.in
    #   task-manager
idna==3.10
    # via
    #   -r requirements.in
    #   requests
packaging==24.2
    # via
    #   -r requirements.in
    #   gunicorn
psycopg2-binary==2.9.10
    # via
    #   -r requirements.in
    #   task-manager
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   task-manager
requests==2.32.3
    # via
    #   -r requirements.in
    #   rollbar
rollbar==1.1.2
    # via
    #   -r requirements.in
    #   task-manager
ruff==0.9.3
    # via -r requirements.in
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   dj-database-url
tzdata==2025.1
    # via
    #   -r requirements.in
    #   django
unicorn==2.1.1
    # via
    #   -r requirements.in
    #   task-manager
urllib3==2.3.0
    # via
    #   -r requirements.in
    #   requests
